requirements.txt
# requirements.txt

# Thư viện cốt lõi
requests>=2.32.0
numpy>=2.2.3
PyYAML>=6.0.2
argparse>=1.4.0
python-dotenv>=0.20.0
pydantic>=1.9.0

# Logging và giao diện
rich>=12.0.0
colorlog>=6.9.0
tqdm>=4.64.0

# Xử lý dữ liệu
sqlite3
pandas>=1.4.0

# Phát triển
pytest>=7.0.0
pytest-cov>=3.0.0
mypy>=0.950
flake8>=4.0.0
black>=22.3.0
isort>=5.10.0

# Tài liệu
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0
